plato class implemented + render in main
