+++ src/chess_game/plato.h
@@ -18,8 +18,8 @@ class Plato
         Plato();
         ~Plato();
         void Draw(VertexArray& va, Camera& cam, Shader& shader, Renderer& renderer);
-        Object caseBlack;
-        Object caseWhite;
+        Object* caseBlack;
+        Object* caseWhite;
         
 
 };
