+++ src/interface_graphique/camera.cpp
@@ -1,7 +1,7 @@
 #include "camera.h"
 #include "renderer.h"
 
-Camera::Camera(float width, float height):position(8,8,30), horizontalAngle(3.14), verticalAngle(0), FoV(45)
+Camera::Camera(float width, float height):position(8,8,10), horizontalAngle(3.14), verticalAngle(0), FoV(45)
 {
     computeMatrices(width, height);
 }
